Moved original crossencoder code to a separate package
