+++ blink/main_dense.py
@@ -7,10 +7,8 @@
 #
 import argparse
 import json
-import sys
 
 from tqdm import tqdm
-import logging
 import torch
 import numpy as np
 from colorama import init
@@ -18,15 +16,14 @@ from termcolor import colored
 
 import blink.ner as NER
 from torch.utils.data import DataLoader, SequentialSampler, TensorDataset
-from blink.biencoder.biencoder import BiEncoderRanker, load_biencoder
-from blink.crossencoder.crossencoder import CrossEncoderRanker, load_crossencoder
+from blink.biencoder.biencoder import load_biencoder
+from blink.crossencoder.crossencoder import load_crossencoder
 from blink.biencoder.data_process import (
     process_mention_data,
-    get_candidate_representation,
 )
 import blink.candidate_ranking.utils as utils
-from blink.crossencoder.train_cross import modify, evaluate
-from blink.crossencoder.data_process import prepare_crossencoder_data
+from blink.crossencoder.original.train_cross import modify, evaluate
+from blink.crossencoder.original.data_process import prepare_crossencoder_data
 from blink.index.faiss_indexer import DenseFlatIndexer, DenseHNSWFlatIndexer
